requirements.txt
numpy~=1.26.2
flask~=3.0.0
Pillow~=10.1.0
keras~=2.15.0
gunicorn==21.2.0